fix(Auth): delete pet with authorization

+++ src/actions/actions.ts
@@ -105,6 +105,13 @@ export async function deletePet(petId: unknown) {
   // mock network time
   await sleep(1000);
 
+  // authentication check
+  const session = await auth();
+  if (!session?.user) {
+    redirect('/login');
+  }
+
+  // validation
   const validatedPetId = petIdSchema.safeParse(petId);
   if (!validatedPetId.success) {
     return {
@@ -112,6 +119,24 @@ export async function deletePet(petId: unknown) {
     };
   }
 
+  // authorization check
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data
+    }
+  })
+  if (!pet) {
+    return {
+      message: 'Pet not found.'
+    }
+  }
+  if (pet.userId !== session.user.id) {
+    return {
+      message: 'Not authorized.',
+    };
+  }
+
+  // db mutation
   try {
     await prisma.pet.delete({
       where: {
